Style nav menu items and make routing work
